+++ modules/swagger-core/src/test/scala/converter/CustomConverterTest.scala
@@ -1,6 +1,6 @@
 package converter
 
-import model._
+import models._
 
 import com.wordnik.swagger.converter._
 import com.wordnik.swagger.model._
