+++ modules/swagger-core/src/test/scala/filter/TestSpecs.scala
@@ -1,6 +1,6 @@
 package filter
 
-import converter.models._
+import models._
 
 import com.wordnik.swagger.converter._
 import com.wordnik.swagger.core.filter._
