+++ modules/swagger-core/src/test/scala/models/ATM.java
@@ -1,4 +1,4 @@
-package converter.models;
+package models;
 
 import java.util.List;
