+++ modules/swagger-core/src/test/scala/models/Currency.java
@@ -1,4 +1,4 @@
-package converter.models;
+package models;
 
 public enum Currency
 {
