+++ modules/swagger-core/src/test/scala/models/JCovariantGetter.java
@@ -1,4 +1,4 @@
-package converter.models;
+package models;
 
 public abstract class JCovariantGetter {
