+++ modules/swagger-core/src/test/scala/models/JsonSchemaModelSerializationTest.scala
@@ -1,4 +1,4 @@
-package model
+package models
 
 import com.wordnik.swagger.model._
