+++ modules/swagger-core/src/test/scala/models/ModelWithAnnotationOnProperty.java
@@ -1,4 +1,4 @@
-package converter.models;
+package models;
 
 import com.wordnik.swagger.annotations.*;
