+++ modules/swagger-core/src/test/scala/models/ModelWithEnumField.java
@@ -1,4 +1,4 @@
-package converter.models;
+package models;
 
 public class ModelWithEnumField {
     public TestEnum enumValue;
