+++ modules/swagger-core/src/test/scala/models/ModelWithEnumProperty.java
@@ -1,4 +1,4 @@
-package converter.models;
+package models;
 
 import javax.xml.bind.*;
 import javax.xml.bind.annotation.*;
