+++ modules/swagger-core/src/test/scala/models/NumericModel.java
@@ -1,3 +1,3 @@
-package model;
+package models;
 
 public class NumericModel extends BaseModel<Integer> {}
