+++ modules/swagger-core/src/test/scala/models/StringModel.java
@@ -1,4 +1,4 @@
-package model;
+package models;
 
 public class StringModel extends BaseModel<String> {
